moved tests to new file
